refactor deploy and use function in test cmd
