+++ cmd/cape/cmd/deploy.go
@@ -1,11 +1,7 @@
 package cmd
 
 import (
-	"archive/zip"
-	"bytes"
 	"fmt"
-	"io"
-	"os"
 	"path/filepath"
 	"strings"
 
@@ -15,21 +11,10 @@ import (
 
 	"github.com/capeprivacy/cli/render"
 	"github.com/capeprivacy/cli/sdk"
-	czip "github.com/capeprivacy/cli/zip"
 )
 
 type DeployResponse struct {
 	ID string `json:"id"`
-}
-
-const storedFunctionMaxBytes = 1_000_000_000
-
-type OversizeFunctionError struct {
-	bytes int64
-}
-
-func (e OversizeFunctionError) Error() string {
-	return fmt.Sprintf("deployment (%d bytes) exceeds size limit of %d bytes", e.bytes, storedFunctionMaxBytes)
 }
 
 // deployCmd represents the request command
@@ -135,70 +120,9 @@ func getName(functionInput string, nameFlag string) string {
 }
 
 func doDeploy(url string, token string, functionInput string, functionName string, insecure bool, pcrSlice []string, public bool) (string, []byte, error) {
-	file, err := os.Open(functionInput)
+	reader, err := sdk.ProcessUserFunction(functionInput)
 	if err != nil {
-		return "", nil, fmt.Errorf("unable to read function directory or file: %w", err)
-	}
-
-	st, err := file.Stat()
-	if err != nil {
-		return "", nil, fmt.Errorf("unable to read function file or directory: %w", err)
-	}
-
-	isZip := false
-	var fileSize int64
-	if st.IsDir() {
-		_, err = file.Readdirnames(1)
-		if err != nil {
-			return "", nil, fmt.Errorf("please pass in a non-empty directory: %w", err)
-		}
-		fileSize, err = dirSize(functionInput)
-		if err != nil {
-			return "", nil, fmt.Errorf("error reading file size: %w", err)
-		}
-	} else {
-		// Check if file ends with ".zip" extension.
-		fileExtension := filepath.Ext(functionInput)
-		if fileExtension != ".zip" {
-			return "", nil, fmt.Errorf("expected argument %s to be a zip file or directory", functionInput)
-		}
-		isZip = true
-		zSize, err := zipSize(functionInput)
-		if err != nil {
-			return "", nil, err
-		}
-
-		fileSize = int64(zSize)
-	}
-
-	log.Debugf("Deployment size: %d bytes", fileSize)
-	if fileSize > storedFunctionMaxBytes {
-		err = OversizeFunctionError{bytes: fileSize}
-		log.Error(err.Error())
 		return "", nil, err
-	}
-
-	err = file.Close()
-	if err != nil {
-		return "", nil, fmt.Errorf("something went wrong: %w", err)
-	}
-
-	var reader io.Reader
-
-	if isZip {
-		f, err := os.Open(functionInput)
-		if err != nil {
-			return "", nil, fmt.Errorf("unable to read function file: %w", err)
-		}
-
-		reader = f
-	} else {
-		buf, err := czip.Create(functionInput)
-		if err != nil {
-			return "", nil, err
-		}
-
-		reader = bytes.NewBuffer(buf)
 	}
 
 	keyReq, err := GetKeyRequest(pcrSlice, token)
@@ -220,30 +144,6 @@ func doDeploy(url string, token string, functionInput string, functionName strin
 	}
 
 	return id, checksum, nil
-}
-
-func dirSize(path string) (int64, error) {
-	var size int64
-	err := filepath.Walk(path, func(_ string, info os.FileInfo, err error) error {
-		if !info.IsDir() {
-			size += info.Size()
-		}
-		return nil
-	})
-	return size, err
-}
-
-func zipSize(path string) (uint64, error) {
-	var size uint64
-	r, err := zip.OpenReader(path)
-	if err != nil {
-		return 0, err
-	}
-	for _, f := range r.File {
-		size += f.UncompressedSize64
-	}
-
-	return size, nil
 }
 
 func getUsername(t string) (string, error) {
