+++ sdk/deploy.go
@@ -1,10 +1,13 @@
 package sdk
 
 import (
+	"archive/zip"
 	"bytes"
 	"crypto/sha256"
 	"fmt"
 	"io"
+	"os"
+	"path/filepath"
 
 	"github.com/gorilla/websocket"
 	log "github.com/sirupsen/logrus"
@@ -16,7 +19,18 @@ import (
 	"github.com/capeprivacy/cli/entities"
 	"github.com/capeprivacy/cli/pcrs"
 	proto "github.com/capeprivacy/cli/protocol"
+	czip "github.com/capeprivacy/cli/zip"
 )
+
+const storedFunctionMaxBytes = 1_000_000_000
+
+type OversizeFunctionError struct {
+	bytes int64
+}
+
+func (e OversizeFunctionError) Error() string {
+	return fmt.Sprintf("deployment (%d bytes) exceeds size limit of %d bytes", e.bytes, storedFunctionMaxBytes)
+}
 
 type protocol interface {
 	WriteStart(request entities.StartRequest) error
@@ -149,4 +163,99 @@ func writeFunction(conn *websocket.Conn, reader io.Reader) error {
 	}
 
 	return nil
+}
+
+// ProcessUserFunction takes a string path and produces a io.Reader to zipped function.
+// It could take both a folder as well as a zip file.
+func ProcessUserFunction(path string) (io.Reader, error) {
+	file, err := os.Open(path)
+	if err != nil {
+		return nil, fmt.Errorf("unable to read function directory or file: %w", err)
+	}
+
+	st, err := file.Stat()
+	if err != nil {
+		return nil, fmt.Errorf("unable to read function file or directory: %w", err)
+	}
+
+	isZip := false
+	var fileSize int64
+	if st.IsDir() {
+		_, err = file.Readdirnames(1)
+		if err != nil {
+			return nil, fmt.Errorf("please pass in a non-empty directory: %w", err)
+		}
+		fileSize, err = dirSize(path)
+		if err != nil {
+			return nil, fmt.Errorf("error reading file size: %w", err)
+		}
+	} else {
+		// Check if file ends with ".zip" extension.
+		fileExtension := filepath.Ext(path)
+		if fileExtension != ".zip" {
+			return nil, fmt.Errorf("expected argument %s to be a zip file or directory", path)
+		}
+		isZip = true
+		zSize, err := zipSize(path)
+		if err != nil {
+			return nil, err
+		}
+
+		fileSize = int64(zSize)
+	}
+
+	log.Debugf("Deployment size: %d bytes", fileSize)
+	if fileSize > storedFunctionMaxBytes {
+		err = OversizeFunctionError{bytes: fileSize}
+		log.Error(err.Error())
+		return nil, err
+	}
+
+	err = file.Close()
+	if err != nil {
+		return nil, fmt.Errorf("something went wrong: %w", err)
+	}
+
+	var reader io.Reader
+
+	if isZip {
+		f, err := os.Open(path)
+		if err != nil {
+			return nil, fmt.Errorf("unable to read function file: %w", err)
+		}
+
+		reader = f
+	} else {
+		buf, err := czip.Create(path)
+		if err != nil {
+			return nil, err
+		}
+
+		reader = bytes.NewBuffer(buf)
+	}
+	return reader, nil
+}
+
+func zipSize(path string) (uint64, error) {
+	var size uint64
+	r, err := zip.OpenReader(path)
+	if err != nil {
+		return 0, err
+	}
+	for _, f := range r.File {
+		size += f.UncompressedSize64
+	}
+
+	return size, nil
+}
+
+func dirSize(path string) (int64, error) {
+	var size int64
+	err := filepath.Walk(path, func(_ string, info os.FileInfo, err error) error {
+		if !info.IsDir() {
+			size += info.Size()
+		}
+		return nil
+	})
+	return size, err
 }
